Refactor some scheduling abilities to Job
Refactor package locations, function names, and variable names
for readability and to implement better access levels to fields
that should not be modified.

Refactor definiiton of a Worker to be a struct that contains a Task
interface. A Task interface is what contains the Run() function.

Adjust test cases for new API.