Refactor API (#1)
Refactor package locations, function names, and variable names
for readability and to implement better access levels to fields
that should not be modified.

Refactor definition of a Worker to be a struct that contains a Task
interface. A Task interface is what contains the Run() function.

Refactor some scheduling logic to the Job definition.

Fix race conditions.